+++ app/(content)/dashboard/layout.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-
+import MainPageBody from "@/components/pagetemplates/mainpagebody/mainpagebody";
 import { motion } from "framer-motion";
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
@@ -16,7 +16,9 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
             damping: 20,
             delay: 0.2,
         }}>
-            {children}
+            <MainPageBody>
+                {children}
+            </MainPageBody>
         </motion.div>
     );
 }
