+++ app/(reach)/about/development/desktop.tsx
@@ -2,8 +2,8 @@
 
 import React, { useState, useRef, useCallback } from 'react';
 import Image from 'next/image';
-import { schoolsArray, jobsArray } from '@/components/professionalComponents/jobsarray';
-import { SchoolBite, JobBite } from '@/components/professionalComponents/proBites';
+import { schoolsArray, jobsArray } from '@/components/pagecomponents/professionalComponents/jobsarray';
+import { SchoolBite, JobBite } from '@/components/pagecomponents/professionalComponents/proBites';
 
 const openInNewTab = (url: string) => {
   const win = window.open(url, '_blank');
