+++ components/dropdown/blogdropdown.tsx
@@ -1,0 +1,90 @@
+'use client'
+
+import { ReactHTMLElement, useState } from "react";
+import PostItemList from "../posts/postlistitem";
+import type { post } from "../../types/post";
+import React from "react";
+import InnerHeader from "@/components/pagetemplates/innerheader/InnerHeader";
+import MainChild from "../pagetemplates/mainchild/mainchild";
+import useMediaQuery from "../listeners/WidthSettings";
+
+const BlogDropdown = ({categoriesForDrop, posts}: {categoriesForDrop: Array<string>; posts: Record<string, post[]>}) => {
+    const isBreakpoint = useMediaQuery(768);
+    const [open, setOpen] = useState(false);
+    const [category, setCategory] = useState('All');
+    const dropdownRef = React.useRef<HTMLDivElement>(null);
+    const buttonRef = React.useRef<HTMLDivElement>(null);
+
+    React.useEffect(() => {
+        const handleOutsideClick = (event: { target: any; }) => {
+            if ((!dropdownRef.current || !dropdownRef.current.contains(event.target as HTMLDivElement)) && 
+                (!buttonRef.current || !buttonRef.current.contains(event.target as HTMLElement))) {
+                if (open) {
+                    setOpen(false);
+                }
+            }
+        };
+        window.addEventListener('mousedown', handleOutsideClick);
+        return () => window.removeEventListener('mousedown', handleOutsideClick);
+    }, [dropdownRef, open, setOpen, buttonRef]);
+
+    return (
+        <>
+            <InnerHeader>
+                <header className="underline text-4xl">
+                    Blog
+                </header>
+                <div ref={buttonRef} className="flex flex-row items-center">
+                    <button id="dropButton" onClick={open ? () => setOpen(false) : () => setOpen(true)} className='flex flex-row justify-between cursor-pointer'>
+                        <p className='flex pr-2'>
+                            Filter:
+                        </p>
+                        <div className='relative text-black rounded'>
+                            {category}
+                        </div>
+                    </button>
+                    <div className='flex items-end'>
+                        <svg
+                            className="h-4"
+                            xmlns="http://www.w3.org/2000/svg"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                            aria-hidden="true"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                    </div>
+                </div>
+                {open && 
+                    <div ref={dropdownRef} className='absolute flex flex-col z-30 right-16 top-32 mt-2 justify-end text-left border border-gray-300 rounded-sm bg-clip-padding bg-slate-800/70 text-white shadow-lg w-32 cursor-pointer'>
+                        {categoriesForDrop.map((category, id) => (
+                            <div key={id} onClick={() => {
+                                setCategory(category)
+                                setOpen(false)
+                            }} className='block px-4 py-2 text-sm text-white hover:bg-slate-800'>
+                                {category}
+                            </div>
+                        ))}
+                    </div>
+                }
+            </InnerHeader>   
+            <MainChild>
+                <div className={`${category === 'All' || !isBreakpoint ? `md:grid md:grid-cols-2 justify-center` : `items-center justify-start`}`}>
+                        {posts !== null && category === 'All' &&
+                            Object.keys(posts).map((category, id) => <PostItemList key={id} category={category} posts={posts[category]} />
+                            )
+                        }
+                        {posts !== null && category !== 'All' &&
+                            <PostItemList category={category} posts={posts[category]} />
+                        }
+                </div>
+            </MainChild>
+            </>
+    );
+    }
+
+export default BlogDropdown;
