+++ components/posts/postlistitem.tsx
@@ -2,12 +2,12 @@ import Link from 'next/link';
 import type { post } from '../../types/post';
 
 export default function PostItemList({ category, posts }: { category: string; posts: post[] }) {
-    const sortedPosts = posts.sort((a, b) => a.date < b.date ? -1 : 1);
+    const sortedPosts = posts.sort((a, b) => a.title < b.title ? -1 : 1);
 
     return (
         <div className='flex flex-col gap-5'>
             <h2 className='text-2xl font-cormorantGaramond underline'>{category}</h2>
-            <div className='flex flex-col gap-2.5 text-base' style={{minHeight: '14vh', maxHeight: '14vh', overflowY: 'auto'}}>
+            <div className='flex flex-col gap-2.5 pb-5 text-base'>
                 {sortedPosts.map((post, id) => (
                     <Link href={`/${post.id}`} key={id} className='text-neutral-900 hover:text-amber-700 transition duration-150'>
                         -{post.title}
