--- components/trackers/taskTrackers.tsx
@@ -1,9 +1,0 @@
-const TaskTracker = () => {
-    return (
-        <div>
-            <h1>Task Tracker</h1>
-        </div>
-    );
-};
-
-export default TaskTracker;
